stat: Prevent the BW and IOPS logging interval from creeping up
With the current FIO version, we're seeing the interval for writing
the BW and IOPS logs increase by 1 msec over time.  The --log_avg_msec
option is used to set the time interval for logging BW and IOPs.

This issue will cause dip ( low ) in IOPS and BW plot vs time because
parsing/plot tool can't get correct sum total IOPS and BW from
multiple jobs when there is 1-2us increment in timestamp.

The issue was introduced in version 3.18 with the following
commit - 31eca64.  And these changes made it slightly better: 0f77d30,
and PR #973 (0e59dd6), but the problem still exists.

This commit also contains the change from PR #982.  I believe we need
both changes to totally resolve the issue of the BW and IOPs logging
interval creep.

Signed-off-by: Jeff Lien <jeff.lien@wdc.com>